[CUBE-7] (feat): for invalid/no-access workspaces show 404 on same url
