+++ src/pages/WorkspaceGuard.vue
@@ -1,0 +1,55 @@
+<template>
+  <Components :is="component" />
+</template>
+
+<script>
+import Workspace from '/pages/Workspace'
+import NotFound from '/pages/NotFound'
+import gql from 'graphql-tag'
+
+export default {
+  name: 'WorkspaceGuard',
+  apollo: {
+    workspace: {
+      client: 'work',
+      query: gql`
+        query currentWorkspace($spaceUrl: String!) {
+          workspace(name: $spaceUrl) {
+            name
+          }
+        }
+      `,
+      variables() {
+        return {
+          spaceUrl: this.$route.params.workspace,
+        }
+      },
+      result({ data }) {
+        this.showWorkspace = data.workspace !== null
+        if (data.workspace)
+          this.$apollo.provider.clients.work.cache.writeData({ data })
+      },
+      error(err) {
+        if (
+          err.message.replace('GraphQL error:', '').trim() === 'NOT_LOGGED_IN'
+        ) {
+          this.$router.push('/login')
+          return
+        } else throw Error(err)
+      },
+    },
+  },
+  components: {
+    Workspace,
+    NotFound,
+  },
+  data: () => ({
+    showWorkspace: false,
+  }),
+  computed: {
+    component() {
+      return this.showWorkspace ? 'Workspace' : 'NotFound'
+    },
+  },
+}
+</script>
